[Autosuggest] Remove spike naming, and some more cleanup of suggestions UI
