+++ src/Components/__stories__/Search.story.tsx
@@ -3,15 +3,8 @@ import React from "react"
 
 import { ContextProvider } from "Artsy/SystemContext"
 import { SearchPreview } from "Components/Search/Previews"
-import { SearchBar } from "Components/Search/SearchBar"
-import { SearchBarSpikeQueryRenderer as SearchBarSpike } from "Components/Search/Spike/SearchBarSpike"
+import { SearchBarQueryRenderer as SearchBar } from "Components/Search/SearchBar"
 import { SearchSuggestionsQueryRenderer as SearchSuggestions } from "Components/Search/Suggestions"
-
-storiesOf("Components/Search/Spike/SearchBarSpike", module).add("Input", () => (
-  <ContextProvider>
-    <SearchBarSpike />
-  </ContextProvider>
-))
 
 storiesOf("Components/Search/SearchBar", module).add("Input", () => (
   <ContextProvider>
